backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
email-validator==2.1.1
neo4j==5.25.0
python-dotenv==1.0.1
starlette==0.38.6
python-multipart==0.0.20
dotenv
